Update 1-virt-manager.md
I just added a small note saying that the password for your windows vm cannot be blank. 
I left it blank at first because I thought it would be easier and then later on it didn't work and it took me a while to realize that this was the reason.